[41] feat : 전역에서 유저 프로필 관리하도록 수정
