+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import "@/app/globals.scss";
 import styles from "@/app/layout.module.scss";
 import Providers from "@/components/Providers";
+import UserProfileProvider from "@/components/UserProfileProvider";
 
 export const metadata: Metadata = {
   title: "투스데이",
@@ -16,7 +17,9 @@ export default function RootLayout({
   return (
     <html lang="ko">
       <body className={styles.root}>
-        <Providers>{children}</Providers>
+        <Providers>
+          <UserProfileProvider>{children}</UserProfileProvider>
+        </Providers>
       </body>
     </html>
   );
