+++ src/components/medical/UserWelcome.tsx
@@ -2,38 +2,9 @@
 import { useUserStore } from "@/stores/user";
 import styles from "./UserWelcome.module.scss";
 import Link from "next/link";
-import { fetchUserProfile } from "@/api/authService";
-import { useQuery } from "react-query";
-import Loading from "@/app/loading";
-import Error from "../error/Error";
 
 const UserWelcome = () => {
   const { userProfile, setUserProfile } = useUserStore();
-
-  const { data, error, isLoading } = useQuery({
-    queryKey: "userProfile",
-    queryFn: () => fetchUserProfile(),
-    onSuccess: (data) => {
-      const profile = {
-        id: data.id,
-        email: data.email,
-        profileImageUrl: data.profileImageUrl,
-        username: data.username
-      };
-      setUserProfile(profile);
-    },
-    onError: (error) => {
-      console.error("Failed to fetch user profile", error);
-    }
-  });
-
-  if (isLoading) {
-    return <Loading />;
-  }
-
-  if (error) {
-    return <Error errorType={"error"} />;
-  }
 
   return (
     <section className={styles.medicalSection}>
